Add legacy album response and update search2Results
The page for search2Results has it returning a list of Child objects in
the albums field but this is incorrect. The legacy API docs use an album
type that is not in our documentation.

Add an entry for the legacy album type and update the search2 response
type to use it.

Signed-off-by: Eric B Munson <eric@munsonfam.org>