Add play queue extensions for using index
The current `savePlayQueue` and `getPlayQueue` rely on a `current` being the id of an item in the queue.
This is problematic if the queue has duplicate items, as it is then not possible to determine which index item to use.

This PR adds extensions `getPlayQueue2` and `savePlayQueue2` that rely on a 1-based `currentIndex` to clearly specify which item to use.
Otherwise, they are identical to `getPlayQueue` and `savePlayQueue`.