Add measurementScheme field to ReplayGain Response
The values presented by the ReplayGain object can be misinterpreted
without knowing the standard used in calculating them. It is my
understanding that while these values are all measured in decibles, the
standards used for calculating them make ReplayGain incompatible with
the ReplayGainV2 and EBUR128, but there may be a conversion rate or
factor between ReplayGainV2 and EBUR128. However, I am still not
entirely cleanr on the relationships between the three standards.
However, a play back application could misinterpret the values leading
to unexpected behavoir.

It is not clear to me if we should specify how a client should interpret
the lack of this field or if that can be left up to individual clients
or users.

See: https://github.com/opensubsonic/open-subsonic-api/discussions/165

Signed-off-by: Eric B Munson <eric@munsonfam.org>